agent/requirements.txt
fastapi
uvicorn[standard]
pydantic-ai[openai]
asyncpg
sse-starlette
python-dotenv
httpx
